refactor: move githooks functionality to commands
